marine_trade_ai/requirements.txt
requests==2.31.0
pandas==2.2.1
google-generativeai==0.5.2
newsapi-python==0.2.7
